remove intermediate interface
