+++ webapi/src/main/scala/org/knora/webapi/messages/util/search/gravsearch/prequery/AbstractPrequeryGenerator.scala
@@ -38,8 +38,49 @@ abstract class AbstractPrequeryGenerator(
   constructClause: ConstructClause,
   typeInspectionResult: GravsearchTypeInspectionResult,
   querySchema: ApiV2Schema
-) extends WhereTransformer
-    with ConstructToSelectTransformer {
+) extends WhereTransformer {
+
+  /**
+   * Returns the columns to be specified in the SELECT query.
+   */
+  def getSelectColumns: Task[Seq[SelectQueryColumn]]
+
+  /**
+   * Returns the variables that the query result rows are grouped by (aggregating rows into one).
+   * Variables returned by the SELECT query must either be present in the GROUP BY statement
+   * or be transformed by an aggregation function in SPARQL.
+   * This method will be called by [[QueryTraverser]] after the whole input query has been traversed.
+   *
+   * @param orderByCriteria the criteria used to sort the query results. They have to be included in the GROUP BY statement, otherwise they are unbound.
+   * @return a list of variables that the result rows are grouped by.
+   */
+  def getGroupBy(orderByCriteria: TransformedOrderBy): Task[Seq[QueryVariable]]
+
+  /**
+   * Returns the criteria, if any, that should be used in the ORDER BY clause of the SELECT query. This method will be called
+   * by [[QueryTraverser]] after the whole input query has been traversed.
+   *
+   * @param inputOrderBy the ORDER BY criteria in the input query.
+   * @return the ORDER BY criteria, if any.
+   */
+  def getOrderBy(inputOrderBy: Seq[OrderCriterion]): Task[TransformedOrderBy]
+
+  /**
+   * Returns the limit representing the maximum amount of result rows returned by the SELECT query.
+   *
+   * @return the LIMIT, if any.
+   */
+  def getLimit: Task[Int]
+
+  /**
+   * Returns the OFFSET to be used in the SELECT query.
+   * Provided the OFFSET submitted in the input query, calculates the actual offset in result rows depending on LIMIT.
+   *
+   * @param inputQueryOffset the OFFSET provided in the input query.
+   * @return the OFFSET.
+   */
+  def getOffset(inputQueryOffset: Long, limit: Int): Task[Long]
+
   protected implicit val stringFormatter: StringFormatter = StringFormatter.getGeneralInstance
 
   // a Set containing all `TypeableEntity` (keys of `typeInspectionResult`) that have already been processed
